# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.10.0
  contact:
    name: tr-engine
    url: https://github.com/LumenPrima/tr-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    tr-engine authenticates **client software, not people**. Every client
    (a dashboard deployment, a script, an upload plugin, a website's
    backend, a phone app) holds its own **API key** and sends it as:

    ```
    Authorization: Bearer tre_<64 hex characters>
    ```

    There are no user accounts, logins, sessions or cookies. How a client
    authenticates its own users, if it has any, is up to the client. The
    operator guide and the client-developer guide are in `docs/auth.md`.

    **Scopes.** A key has one of `listen`, `edit` or `admin`, optionally
    plus `upload`, or only `upload`:

    | Scope | Grants | Implies |
    |---|---|---|
    | `listen` | Read radio data, stream events and audio, mint tickets | |
    | `edit` | Change radio metadata (tags, transcription corrections and reviews, unit-tag suggestions) | `listen` |
    | `admin` | Everything else: keys, anonymous access, identity edits, merges, maintenance, storage, SQL query, pages, imports, console logs, audit log | `edit`, `listen` |
...

...
    what applies to it.

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /admin/systems/merge:
    post:
      operationId: mergeSystems
      x-scope: admin
      x-restricted: deny
      summary: Merge two systems
      description: |
        Merges all data from `source_id` into `target_id`, then deletes
        the source system. Use this to fix fragmentation caused by
        instance_id or sys_name changes, or to consolidate an unknown
        system after identification.

        **What happens:**
        - All calls under source are repointed to target system_id
          (call_ids are database PKs and do not change)
        - Talkgroups: source talkgroups are merged into target. If both
          have the same tgid, target metadata is kept and call counts are
          combined. Source-only talkgroups are moved to target.
        - Units: same merge logic as talkgroups
        - Unit events: repointed to target system_id
        - Call groups may be consolidated (recordings that now share
          system_id + tgid + start_time get grouped)
        - Source system is deleted

        - Restrictions on API keys and the anonymous access policy that
          name the source system (in `systems`, `talkgroups` or
          `exclude_talkgroups`) are rewritten to the target in the same
          transaction, so a merge never widens access. Tickets whose
          narrowing names the source system stop verifying
          (`invalid_ticket`); clients mint new ones.

        **Implications:**
        - Call IDs (database PKs) are stable — bookmarks don't break
        - Talkgroup and unit composite IDs change (old system_id prefix
          becomes invalid, new system_id prefix is correct)
        - This operation is **not reversible**

        After merging, update the trunk-recorder config so future MQTT
        messages use the target system's `(instance_id, short_name)`.
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemMergeRequest"
      responses:
        "200":
          description: Merge completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemMergeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Source or target system not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Merge would cause data conflicts that cannot be auto-resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/maintenance:
    get:
      operationId: getMaintenanceStatus
      x-scope: admin
      x-restricted: deny
      summary: Get maintenance status and config
      description: |
        Returns the current maintenance configuration (retention periods)
        and the results of the last maintenance run. Maintenance runs
        automatically every 24 hours on startup.
      tags: [admin]
      responses:
        "200":
          description: Maintenance status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: runMaintenance
      x-scope: admin
      x-restricted: deny
      summary: Trigger immediate maintenance run
      description: |
        Triggers an immediate maintenance run (partition creation,
        decimation, data purging). Returns the results when complete.
        Returns 409 if a maintenance run is already in progress.
      tags: [admin]
      responses:
        "200":
          description: Maintenance run completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRun"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Maintenance already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/transcribe-backfill:
    post:
      operationId: submitTranscribeBackfill
      x-scope: admin
      x-restricted: deny
      summary: Submit transcription backfill job
      description: |
        Queues a background job to find untranscribed calls matching the
        provided filters and drip-feed them into the transcription queue.
        Jobs are processed sequentially — if a job is already active, the
        new job is queued. All filter fields are optional; omitting all
        filters backfills every untranscribed call.

        A job makes a single pass over the matching calls, newest first, and
        hands each call to the transcription queue at most once. A call whose
        transcription then fails (for example, because its audio file is
        missing) stays untranscribed and is not retried by the same job; a
        later job will pick it up again.
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackfillFilters"
      responses:
        "202":
          description: Backfill job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: integer
                    description: Unique identifier for this backfill job
                    example: 1
                  position:
                    type: integer
                    description: Position in the queue (0 = active, 1+ = queued)
                    example: 0
                  total:
                    type: integer
                    description: |
                      Number of untranscribed calls matching filters at
                      submission. Recounted when the job starts.
                    example: 342
                  filters:
                    $ref: "#/components/schemas/BackfillFilters"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Transcription not configured or pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: getTranscribeBackfillStatus
      x-scope: admin
      x-restricted: deny
      summary: Get transcription backfill status
      description: |
        Returns the active and queued backfill jobs. The `active` field
        is null when no job is currently running. The `queued` array
        contains jobs waiting to start, in submission order.
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Backfill status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackfillStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: cancelAllTranscribeBackfill
      x-scope: admin
      x-restricted: deny
      summary: Cancel all backfill jobs
      description: Cancels the active backfill job (if any) and all queued jobs.
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All backfill jobs cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancelled:
                    type: boolean
                    example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No backfill jobs found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/transcribe-backfill/{id}:
    delete:
      operationId: cancelTranscribeBackfill
      x-scope: admin
      x-restricted: deny
      summary: Cancel a specific backfill job
      description: |
        Cancels a specific backfill job by ID. If the job is currently
        active, it is stopped. If it is queued, it is removed from the queue.
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Backfill job ID
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Backfill job cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancelled:
                    type: boolean
                    example: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/maintenance/config:
    put:
      operationId: setMaintenanceConfig
      x-scope: admin
      x-restricted: deny
      summary: Update a retention setting
      description: |
        Updates a single retention setting. The value is persisted to the
        `config_overrides` table and applied immediately to the live pipeline.
        Returns 409 if the key is locked by an environment variable.
        Requires the `admin` scope.
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetentionConfigUpdate"
      responses:
        "200":
          description: Retention setting updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    example: "retention_console_logs"
                  value:
                    type: string
                    example: "48h0m0s"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Key is locked by environment variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/maintenance/config/{key}:
    delete:
      operationId: deleteMaintenanceConfig
      x-scope: admin
      x-restricted: deny
      summary: Reset a retention setting to default
      description: |
        Removes a DB-stored retention override and resets the setting to its
        environment variable value (if set) or coded default. Returns 409 if
        the key is locked by an environment variable.
        Requires the `admin` scope.
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          description: Retention config key (e.g. retention_console_logs)
          schema:
            type: string
            example: "retention_console_logs"
      responses:
        "200":
          description: Retention override removed, reset to default
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    example: "retention_console_logs"
                  status:
                    type: string
                    example: "reset to default"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Key is locked by environment variable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Pipeline not running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: tre_<64 hex characters>
      description: |
        An API key in the `Authorization` header:
        `Authorization: Bearer tre_...`. Keys are created by an admin
        (`POST /keys` or `tr-engine keys create`). Keys are never accepted
        from the URL.
    ticketAuth:
      type: apiKey
      in: query
      name: ticket
      description: |
        A short-lived, listen-only ticket from `POST /tickets`, passed as
        `?ticket=trt_...`. Accepted only by `GET`/`HEAD` on
        `/events/stream`, `/audio/live` and `/calls/{id}/audio`, for
        browser APIs that cannot send headers (`EventSource`,
        `<audio src>`, `WebSocket`). Takes precedence over an
        `Authorization` header on those operations.

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: |
        No usable credential. `code` is `key_required` (no credential, and
        the anonymous access policy doesn't allow this operation),
        `invalid_key` (unknown, revoked or expired key) or `invalid_ticket`
        (malformed, expired or orphaned ticket).
      headers:
        WWW-Authenticate:
          description: Always `Bearer realm="tr-engine"`
          schema:
            type: string
            example: Bearer realm="tr-engine"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            key_required:
              value:
                code: key_required
                error: an API key is required
            invalid_key:
              value:
                code: invalid_key
                error: API key has been revoked
    Forbidden:
      description: |
        The credential is valid but not allowed here. `code` is
        `insufficient_scope` (the message names the scope the operation
        needs) or `restricted_credential` (the credential is restricted to
        some systems or talkgroups, and this operation can't enforce that).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            insufficient_scope:
              value:
                code: insufficient_scope
                error: this operation needs the edit scope
            restricted_credential:
              value:
                code: restricted_credential
                error: this operation is not available to restricted credentials
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
            - key_required
            - invalid_key
            - invalid_ticket
            - insufficient_scope
            - restricted_credential
          description: |
            Machine-readable error code for programmatic handling. Auth codes:
            - `key_required` (401): no credential, and the anonymous access
              policy doesn't allow this operation
            - `invalid_key` (401): unknown, revoked or expired key
            - `invalid_ticket` (401): malformed, expired or orphaned ticket,
              or one that references a merged-away system
            - `insufficient_scope` (403): the key lacks the needed scope
            - `restricted_credential` (403): the credential is restricted and
              this operation can't enforce that
            - `service_unavailable` (503): also when the credential lookup
              failed
            - `forbidden` (403): from the auth layer, only for a route with no
              auth policy (a server bug)
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...
    SystemMergeRequest:
      type: object
      required: [source_id, target_id]
      properties:
        source_id:
          type: integer
          description: System to merge FROM (will be deleted after merge)
          example: 3
        target_id:
          type: integer
          description: System to merge INTO (keeps its system_id)
          example: 1
        update_target_metadata:
          type: boolean
          description: |
            If true, copy short_name/sysid/wacn/nac from source to target
            (useful when the source has better metadata, e.g., was identified
            after the target was created as "unknown").
          default: false
    SystemMergeResponse:
      type: object
      properties:
        target_id:
          type: integer
          description: The surviving system_id
          example: 1
        source_id:
          type: integer
          description: The deleted system_id
          example: 3
        calls_moved:
          type: integer
          description: Number of call records repointed
          example: 5200
        talkgroups_moved:
          type: integer
          description: Talkgroups moved from source (source-only)
          example: 12
        talkgroups_merged:
          type: integer
          description: Talkgroups that existed in both (stats combined)
          example: 45
        units_moved:
          type: integer
          example: 30
        units_merged:
          type: integer
          example: 150
        events_moved:
          type: integer
          description: Unit events repointed
          example: 25000
...
    MaintenanceStatus:
      type: object
      properties:
        config:
          $ref: "#/components/schemas/MaintenanceConfig"
        last_run:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/MaintenanceRun"
          description: Results of the last maintenance run. Null if no run has completed since startup.
...
    MaintenanceRun:
      type: object
      description: Results of a single maintenance run.
      properties:
        started_at:
          type: string
          format: date-time
          description: When the maintenance run started
        duration_ms:
          type: integer
          description: Total run duration in milliseconds
          example: 1234
        partitions_created:
          type: integer
          description: Number of new partitions created
          example: 0
        partitions_dropped:
          type: array
          items:
            type: string
          description: Names of old partitions that were dropped
          example: ["mqtt_raw_messages_w2026_07"]
        decimation:
          type: object
          description: Per-table decimation results
          additionalProperties:
            $ref: "#/components/schemas/DecimationResult"
          example:
            recorder_snapshots: { phase1_deleted: 500, phase2_deleted: 120 }
            decode_rates: { phase1_deleted: 300, phase2_deleted: 80 }
        purged:
          type: object
          description: Per-table purge row counts
          additionalProperties:
            type: integer
          example:
            console_messages: 42
            plugin_statuses: 0
            call_active_checkpoints: 15
            stale_calls: 3
            orphan_call_groups: 1
    DecimationResult:
      type: object
      properties:
        phase1_deleted:
          type: integer
          description: "Rows deleted in phase 1 (1 week – 1 month old: keep 1 per minute)"
          example: 500
        phase2_deleted:
          type: integer
          description: "Rows deleted in phase 2 (>1 month old: keep 1 per hour)"
          example: 120
    BackfillFilters:
      type: object
      description: Optional filters for a transcription backfill job. All fields are optional — omitting all filters targets every untranscribed call.
      properties:
        system_id:
          type: integer
          description: Restrict to calls from this system
          example: 1
        tgids:
          type: array
          items:
            type: integer
          description: Restrict to calls on these talkgroup IDs
          example: [9178, 9112]
        start_time:
          type: string
          format: date-time
          description: Only include calls after this time (RFC 3339)
        end_time:
          type: string
          format: date-time
          description: Only include calls before this time (RFC 3339)
...
    BackfillStatus:
      type: object
      description: Current state of the transcription backfill system.
      properties:
        active:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/BackfillJob"
          description: Currently running backfill job, or null if idle
        queued:
          type: array
          items:
            $ref: "#/components/schemas/BackfillJob"
          description: Jobs waiting to start, in submission order
...
    RetentionConfigUpdate:
      type: object
      description: Request body for updating a retention setting.
      required: [key, value]
      properties:
        key:
          type: string
          description: Retention config key
          enum:
            - retention_raw_messages
            - retention_console_logs
            - retention_plugin_status
            - retention_trunking_messages
            - retention_checkpoints
            - retention_stale_calls
          example: "retention_console_logs"
        value:
          type: string
          description: "Go duration string (e.g. 48h, 720h, 7d)"
          example: "48h"
...